feat: centralized login drawer
