+++ src/locales/en/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"6YtxFj\":[\"Name\"],\"S5Bu+B\":[\"Enter your name here\"],\"7sNhEz\":[\"Username\"],\"zmL8Lp\":[\"Enter your username here\"],\"E+oKCO\":[\"Bio\"],\"O8cFJC\":[\"Enter your bio here\"],\"yjVihA\":[\"Profile Image\"],\"XpjWSd\":[\"Customized Link\"],\"dn+x7E\":[\"Add URL\"],\"MPPZ54\":[\"Add account\"],\"vERlcd\":[\"Profile\"],\"ZCYobq\":[\"Links\"],\"O2GxA2\":[\"Edit profile and account\"],\"FxVG/l\":[\"Copied to clipboard\"],\"he3ygx\":[\"Copy\"],\"ZopSbj\":[\"Recently Updated\"],\"4Q5LRX\":[\"Listing since\"],\"zZOZj7\":[\"Likes\"],\"tuhghE\":[\"Edit list\"],\"fOKqw/\":[\"Add an idea\"],\"VSniCT\":[\"List Cover\"],\"ZXeH4g\":[\"Idea List\"],\"MCLjtL\":[\"Edit list cover\"],\"VsDFHC\":[\"Your ideas live here. Create one!\"],\"IrqaKp\":[\"Tap an idea to edit. Drag the lines on the left to reorder.\"],\"rhRxkK\":[\"Edit List\"],\"V22EiC\":[\"Save New Order\"],\"b3p7OM\":[\"Deleting the list title will also erase all ideas!\"],\"KMfg9q\":[\"Permanently delete the entire list and all its ideas. This action cannot be undone!\"],\"THIBos\":[\"Delete List and All Ideas\"],\"dEgA5A\":[\"Cancel\"],\"P0rbCt\":[\"Cover Image\"],\"gsjH+v\":[\"Hey, the title can't be left empty!\"],\"E93Wuj\":[\"Every list needs a title, so fill it in!\"],\"dBg1Gg\":[\"List title is too long!\"],\"J/tFrO\":[\"Please keep it under \",[\"TITLE_MAX_LENGTH\"],\" characters.\"],\"l6w5x5\":[\"Description is too long!\"],\"F6+KZb\":[\"Please keep it under \",[\"DESC_MAX_LENGTH\"],\" characters.\"],\"JgzDWT\":[\"This is the title of your list\"],\"w/K9b6\":[\"Describe what this title is about\"],\"KHDCQm\":[\"Link a page\"],\"nllOOr\":[\"Edit List Topic\"],\"gbxG1z\":[\"List Topic\"],\"FVbPOI\":[\"Choose a topic that vibes with your List.\"],\"hXzOVo\":[\"Next\"],\"edl1Bx\":[\"Your edits will be lost if you cancel!\"],\"5ttevj\":[\"If you cancel, everything you've entered will be lost.\"],\"PsVVX9\":[\"Cancel Editing\"],\"hoaavJ\":[\"Continue Editing\"],\"AOvWO2\":[\"Create Idea List\"],\"/IX/7x\":[\"Other\"],\"rBfbli\":[\"Lifestyle\"],\"lWxAUo\":[\"Food & Drink\"],\"wFuUuH\":[\"Culture\"],\"lj2WeU\":[\"Travel\"],\"x7ZLZ3\":[\"Entertainment\"],\"kAgdQQ\":[\"Tech & Digital\"],\"FgFxl9\":[\"Personal Growth\"],\"bLbRrJ\":[\"Health & Fitness\"],\"uyJsf6\":[\"About\"],\"4Y5H+g\":[\"Terms\"],\"rjGI/Q\":[\"Privacy\"],\"ab3OdJ\":[\"Relist works best on mobile. Use a mobile device for the best experience.\"],\"/fOrgx\":[\"<0>Your lists say more than posts ever could.</0><1><2>Next wave of content</2><3/></1>\"],\"sfb9Ng\":[\"<0>My life in my lists</0><1>Invitation only. Ready to join?</1>\"],\"cpE88+\":[\"Create your account\"],\"5lWFkC\":[\"Sign in\"],\"NuKR0h\":[\"Others\"],\"jAN+Se\":[\"Lists that vibe right\"],\"8YtR7h\":[\"Lists that taste like wow\"],\"DLW/m0\":[\"Lists that show your colors\"],\"o2i9EJ\":[\"Lists that go to places\"],\"MlvbpW\":[\"Lists that stream good times\"],\"2GbPea\":[\"Lists that tap into the future\"],\"3GYIUI\":[\"Lists that build better you\"],\"YeTwPt\":[\"Lists that make you sweat\"],\"IUd4br\":[\"Lists that just get it\"],\"EiP0f+\":[\"<0>Recently updated</0>\"],\"V1fa9u\":[\"See more\"],\"Zs3bDD\":[\"Fresh lists from our pocket\"],\"ZJ0qYg\":[\"Updated weekly\"],\"rP2JNC\":[\"Oops something is wrong!\"],\"hYyEgm\":[\"<0>Relist works best on mobile.</0><1>Use a mobile device for the best experience.</1>\"],\"tWA9nu\":[\"Idea\"],\"AJoiN9\":[\"Are you sure you want to delete this idea?\"],\"KCgg1o\":[\"Once deleted, this idea cannot be recovered!\"],\"N6GBcC\":[\"Confirm Delete\"],\"3YNSm9\":[\"Every idea needs a title, so fill it in!\"],\"ClX0MT\":[\"Idea title is too long!\"],\"Z/FWct\":[\"This is the title of your idea\"],\"G4krEw\":[\"Describe what this idea is about\"],\"gMn81o\":[\"Title is saved, but your idea edits will be lost! Are you sure you want to cancel?\"],\"jf150A\":[\"Edit Idea\"],\"D9WD3n\":[\"New Idea\"],\"DPfwMq\":[\"Done\"],\"0ESV18\":[\"Life Style\"],\"h16FyT\":[\"Lists\"],\"VQmsFR\":[\"Select your language!\"],\"rtZ04n\":[\"Your preferred language for a better experience.\"],\"if5Kxg\":[\"Taiwan\"],\"Mqy/Zy\":[\"United States\"],\"E6OqPk\":[\"Select your location!\"],\"FsmHKW\":[\"Your location for a personalized experience.\"],\"iZJxvU\":[\"Preference\"],\"AXTJAW\":[\"Select your preferred language\"],\"FFy8Oz\":[\"Select your location\"],\"DNeL+Q\":[\"About Relist\"],\"8jGB5r\":[\"Quick start guide\"],\"RJ5xTx\":[\"Check out Relist\"],\"3fMxeH\":[\"Follow Relist Threads\"],\"YbVeAM\":[\"Join Relist Discord\"],\"bucbEL\":[\"Submit feedback or report content\"],\"2OUiOQ\":[\"About Privacy Policy and Terms of Use\"],\"Bhgd0l\":[\"Contact us\"],\"n1ekoW\":[\"Sign In\"],\"vzX5FB\":[\"Delete Account\"],\"bHYIks\":[\"Sign Out\"],\"VdL8Lz\":[\"Welcome to the Relist community!\"],\"zaNICH\":[\"This is your space to drop ideas and build your pocket list. Tell us what you think and how it feels. Your voice helps shape what comes next.\"],\"OEtKMx\":[\"Tell us what you think\"],\"EjIvar\":[\"Setting Center\"],\"k63iCK\":[\"The login session is expired, please login again\"],\"y6sq5j\":[\"Following\"],\"MKEPCY\":[\"Follow\"],\"x5LEuB\":[\"Followers\"],\"nqCai9\":[\"Links\"],\"kBG60d\":[\"Looks like this page is a bit empty.\"],\"5cg++T\":[\"Follow and check back soon for surprise lists!\"],\"BuijXr\":[\"<0>\",[\"0\"],\"</0>\"],\"FATCWC\":[\"New to Relist? Get started!\"],\"TGUAu2\":[\"By continuing, you agree to our <0>Terms of Service</0>, <1>Privacy Policy</1>. You confirm you're 13+.\"],\"v0GDxO\":[\"Enter your text here\"],\"tfDRzk\":[\"Save\"],\"W9FRBT\":[\"Like\"],\"yVFoGP\":[\"Create List\"],\"6eQ0FA\":[\"Image must be at least 150x150px.\"],\"s6QF9r\":[\"Choose your image\"],\"aHlf/t\":[\"500x500px, JPG or PNG, max 4MB\"],\"xHDYLK\":[\"Upload a cover\"],\"UrzYP4\":[\"Right now, only invited users can log in\"],\"DGilck\":[\"Already got your invite? Jump in and apply now!\"],\"VaEPrX\":[\"Join Relist and share your lists.\"],\"Vm9srO\":[\"Currently open only to creator account applications.\"],\"e+RpCP\":[\"Sign up\"],\"/evPCn\":[\"Creator Account Login\"],\"JX4ezt\":[\"Can non-creators use Relist?\"],\"/q7UJE\":[\"Real impact is simple\"],\"JOfh8O\":[\"Turn your picks into a following\"],\"2doGYd\":[\"Inspogirlie\"],\"YzMqiP\":[\"Must-Haves for dorm life\"],\"Jrcjjh\":[\"Cozy Bedding:\"],\"olcKig\":[\"A plush comforter and extra pillows for a comfy night's sleep.\"],\"GPzBEq\":[\"Desk Essentials\"],\"CmxV+G\":[\"Target is probably your bf.\"],\"LKwNpo\":[\"Portable Speaker\"],\"GHHXT0\":[\"For study breaks or casual dorm hangouts\"],\"V/D+9x\":[\"List Lab\"],\"vQ65+r\":[\"Stop-by-Stop Café Journey\"],\"Zzo4uk\":[\"8:00 AM at Sunrise Café\"],\"9DUK98\":[\"Start your day with a creamy latte and a freshly baked croissant\"],\"c+8j4b\":[\"12:30 PM at Artisan Coffee House\"],\"eVVHMA\":[\"Enjoy a flat white with a hearty avocado toast.\"],\"tEpROq\":[\"3:00 PM at Brew & Bloom\"],\"LrHd3m\":[\"Sip on a cold brew while indulging in a lavender scone\"],\"mPs/O3\":[\"Campus People\"],\"x3KB6d\":[\"Notes for Campus Fest\"],\"UQjid/\":[\"Essentials\"],\"nYR44Q\":[\"Sunscreen, water bottle, and comfy shoes.\"],\"LeaWVj\":[\"Extras\"],\"uqajLC\":[\"A light jacket for when it gets chilly.\"],\"F7HoVk\":[\"Pro tip:\"],\"O1b0Cr\":[\"Bring a blanket to claim your spot early.\"],\"4oJHLS\":[\"Travel Lover\"],\"OWjR7r\":[\"Travel budgeting checklist\"],\"Ir1CN/\":[\"Set your total budget\"],\"c2eiO8\":[\"Break it down into categories: flights, accommodation, food, activities, and extras.\"],\"SZ6k2f\":[\"Hunt for Deals\"],\"TLPUAN\":[\"Use flight comparison tools like Skyscanner or Google Flights\"],\"QdUQ+N\":[\"Transportation Hacks\"],\"B1JN8A\":[\"Book accommodations early through budget-friendly platforms like Hostelworld or Airbnb.\"],\"qEIbYn\":[\"Gossip Pie\"],\"jUlzvp\":[\"Binge based on your mood\"],\"0U3C98\":[\"Relaxing: Moana\"],\"CXBJE9\":[\"Moana / A heartwarming and adventurous animated film with stunning visuals and beautiful music that will soothe your soul.\"],\"c3jXdn\":[\"Tense: Tenet\"],\"5Al7Bs\":[\"Directed by Christopher Nolan, this mind-bending action film with time inversion will keep you on the edge of your seat.\"],\"mRxtxe\":[\"Fun: Dumb and Dumber\"],\"xGoWE/\":[\"A classic comedy about two clueless friends on a road trip full of ridiculous and laugh-out-loud situations.\"],\"SMFKok\":[\"iTech\"],\"PdTFHs\":[\"How I curate my digital space\"],\"624Mtx\":[\"Notion\"],\"ba9xfk\":[\"Manage your thoughts and ideas in one single place.\"],\"w9Xmw/\":[\"Google Drive\"],\"Q2veSe\":[\"The OG and it's so essential.\"],\"yOaaWk\":[\"ChatGPT\"],\"DfSZt9\":[\"Greatly improved my work!\"],\"kPUYC2\":[\"Good Habbits\"],\"gS0YFj\":[\"Atomic Habits - My thoughts\"],\"QBV5Dh\":[\"Start Small to Achieve Big\"],\"x3x8gi\":[\"Pressure is the biggest enemy. Don't let it control your momentum.\"],\"nkJmeJ\":[\"Build Systems, Not Just Goals.\"],\"Pzs/Wu\":[\"This reminds me that building team is more important than the actual product.\"],\"hVijR6\":[\"Embrace Your New Identity\"],\"n+De4p\":[\"I need to work on this 100%\"],\"iEfYGf\":[\"Pica Pica\"],\"sPzKc0\":[\"My best protein shake recipe\"],\"rkUWz9\":[\"Step 1: Gather Your Ingredients\"],\"BLAQSh\":[\"1 scoop of your favorite protein powder (vanilla or chocolate works great!)\"],\"HJwOp1\":[\"Step 2: Blend the Base\"],\"8VH0aE\":[\"Add the almond milk, protein powder, and banana to your blender.\"],\"OOBT5G\":[\"Step 3: Add the Extras\"],\"XyYukS\":[\"Drop in the peanut butter, chia seeds, or flaxseeds for added nutrients.\"],\"G9hWXm\":[\"Jacob PHD\"],\"LGdu/V\":[\"Reasons I swipe left\"],\"/DEtrq\":[\"The Fish Pic\"],\"K5eTbU\":[\"Congrats on catching that trout, but it's not winning you dates.\"],\"IiU328\":[\"No Bio, No Effort\"],\"3KK42j\":[\"If you can't write a sentence, how are you going to hold a conversation?\"],\"jSx6Fj\":[\"Mirror Selfies with a Dirty Room\"],\"LZQmU/\":[\"If you can't clean your room, what's your life looking like?\"],\"LB7eBr\":[\"Tutorials\"],\"/lDBHm\":[\"FAQ\"],\"pGWSZE\":[\"Join the Team\"],\"XyA7q/\":[\"Press kit\"],\"YcfUZ9\":[\"Contact Us\"],\"hKIgWM\":[\"Collection Notice\"],\"4VQIM2\":[\"Login Error - Creator Account Not Found\"],\"ie/LhK\":[\"Only approved and verified “Creator Accounts” can log in and create lists. If you have any questions, please contact us.\"],\"bJ55aX\":[\"Please login again\"],\"sYV3p0\":[\"Please login to do this action\"],\"Pdjj9s\":[\"The app is outdated, please login again\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"6YtxFj\":[\"Name\"],\"S5Bu+B\":[\"Enter your name here\"],\"7sNhEz\":[\"Username\"],\"zmL8Lp\":[\"Enter your username here\"],\"E+oKCO\":[\"Bio\"],\"O8cFJC\":[\"Enter your bio here\"],\"yjVihA\":[\"Profile Image\"],\"XpjWSd\":[\"Customized Link\"],\"dn+x7E\":[\"Add URL\"],\"MPPZ54\":[\"Add account\"],\"vERlcd\":[\"Profile\"],\"ZCYobq\":[\"Links\"],\"O2GxA2\":[\"Edit profile and account\"],\"FxVG/l\":[\"Copied to clipboard\"],\"he3ygx\":[\"Copy\"],\"ZopSbj\":[\"Recently Updated\"],\"4Q5LRX\":[\"Listing since\"],\"zZOZj7\":[\"Likes\"],\"tuhghE\":[\"Edit list\"],\"fOKqw/\":[\"Add an idea\"],\"VSniCT\":[\"List Cover\"],\"ZXeH4g\":[\"Idea List\"],\"MCLjtL\":[\"Edit list cover\"],\"VsDFHC\":[\"Your ideas live here. Create one!\"],\"IrqaKp\":[\"Tap an idea to edit. Drag the lines on the left to reorder.\"],\"rhRxkK\":[\"Edit List\"],\"V22EiC\":[\"Save New Order\"],\"b3p7OM\":[\"Deleting the list title will also erase all ideas!\"],\"KMfg9q\":[\"Permanently delete the entire list and all its ideas. This action cannot be undone!\"],\"THIBos\":[\"Delete List and All Ideas\"],\"dEgA5A\":[\"Cancel\"],\"P0rbCt\":[\"Cover Image\"],\"gsjH+v\":[\"Hey, the title can't be left empty!\"],\"E93Wuj\":[\"Every list needs a title, so fill it in!\"],\"dBg1Gg\":[\"List title is too long!\"],\"J/tFrO\":[\"Please keep it under \",[\"TITLE_MAX_LENGTH\"],\" characters.\"],\"l6w5x5\":[\"Description is too long!\"],\"F6+KZb\":[\"Please keep it under \",[\"DESC_MAX_LENGTH\"],\" characters.\"],\"JgzDWT\":[\"This is the title of your list\"],\"w/K9b6\":[\"Describe what this title is about\"],\"KHDCQm\":[\"Link a page\"],\"nllOOr\":[\"Edit List Topic\"],\"gbxG1z\":[\"List Topic\"],\"FVbPOI\":[\"Choose a topic that vibes with your List.\"],\"hXzOVo\":[\"Next\"],\"edl1Bx\":[\"Your edits will be lost if you cancel!\"],\"5ttevj\":[\"If you cancel, everything you've entered will be lost.\"],\"PsVVX9\":[\"Cancel Editing\"],\"hoaavJ\":[\"Continue Editing\"],\"AOvWO2\":[\"Create Idea List\"],\"/IX/7x\":[\"Other\"],\"rBfbli\":[\"Lifestyle\"],\"lWxAUo\":[\"Food & Drink\"],\"wFuUuH\":[\"Culture\"],\"lj2WeU\":[\"Travel\"],\"x7ZLZ3\":[\"Entertainment\"],\"kAgdQQ\":[\"Tech & Digital\"],\"FgFxl9\":[\"Personal Growth\"],\"bLbRrJ\":[\"Health & Fitness\"],\"uyJsf6\":[\"About\"],\"4Y5H+g\":[\"Terms\"],\"rjGI/Q\":[\"Privacy\"],\"ab3OdJ\":[\"Relist works best on mobile. Use a mobile device for the best experience.\"],\"/fOrgx\":[\"<0>Your lists say more than posts ever could.</0><1><2>Next wave of content</2><3/></1>\"],\"sfb9Ng\":[\"<0>My life in my lists</0><1>Invitation only. Ready to join?</1>\"],\"cpE88+\":[\"Create your account\"],\"5lWFkC\":[\"Sign in\"],\"NuKR0h\":[\"Others\"],\"jAN+Se\":[\"Lists that vibe right\"],\"8YtR7h\":[\"Lists that taste like wow\"],\"DLW/m0\":[\"Lists that show your colors\"],\"o2i9EJ\":[\"Lists that go to places\"],\"MlvbpW\":[\"Lists that stream good times\"],\"2GbPea\":[\"Lists that tap into the future\"],\"3GYIUI\":[\"Lists that build better you\"],\"YeTwPt\":[\"Lists that make you sweat\"],\"IUd4br\":[\"Lists that just get it\"],\"EiP0f+\":[\"<0>Recently updated</0>\"],\"V1fa9u\":[\"See more\"],\"Zs3bDD\":[\"Fresh lists from our pocket\"],\"ZJ0qYg\":[\"Updated weekly\"],\"rP2JNC\":[\"Oops something is wrong!\"],\"hYyEgm\":[\"<0>Relist works best on mobile.</0><1>Use a mobile device for the best experience.</1>\"],\"tWA9nu\":[\"Idea\"],\"AJoiN9\":[\"Are you sure you want to delete this idea?\"],\"KCgg1o\":[\"Once deleted, this idea cannot be recovered!\"],\"N6GBcC\":[\"Confirm Delete\"],\"3YNSm9\":[\"Every idea needs a title, so fill it in!\"],\"ClX0MT\":[\"Idea title is too long!\"],\"Z/FWct\":[\"This is the title of your idea\"],\"G4krEw\":[\"Describe what this idea is about\"],\"gMn81o\":[\"Title is saved, but your idea edits will be lost! Are you sure you want to cancel?\"],\"jf150A\":[\"Edit Idea\"],\"D9WD3n\":[\"New Idea\"],\"DPfwMq\":[\"Done\"],\"0ESV18\":[\"Life Style\"],\"h16FyT\":[\"Lists\"],\"VQmsFR\":[\"Select your language!\"],\"rtZ04n\":[\"Your preferred language for a better experience.\"],\"if5Kxg\":[\"Taiwan\"],\"Mqy/Zy\":[\"United States\"],\"E6OqPk\":[\"Select your location!\"],\"FsmHKW\":[\"Your location for a personalized experience.\"],\"iZJxvU\":[\"Preference\"],\"AXTJAW\":[\"Select your preferred language\"],\"FFy8Oz\":[\"Select your location\"],\"DNeL+Q\":[\"About Relist\"],\"8jGB5r\":[\"Quick start guide\"],\"RJ5xTx\":[\"Check out Relist\"],\"3fMxeH\":[\"Follow Relist Threads\"],\"YbVeAM\":[\"Join Relist Discord\"],\"bucbEL\":[\"Submit feedback or report content\"],\"2OUiOQ\":[\"About Privacy Policy and Terms of Use\"],\"Bhgd0l\":[\"Contact us\"],\"n1ekoW\":[\"Sign In\"],\"vzX5FB\":[\"Delete Account\"],\"bHYIks\":[\"Sign Out\"],\"VdL8Lz\":[\"Welcome to the Relist community!\"],\"zaNICH\":[\"This is your space to drop ideas and build your pocket list. Tell us what you think and how it feels. Your voice helps shape what comes next.\"],\"OEtKMx\":[\"Tell us what you think\"],\"EjIvar\":[\"Setting Center\"],\"k63iCK\":[\"The login session is expired, please login again\"],\"y6sq5j\":[\"Following\"],\"MKEPCY\":[\"Follow\"],\"x5LEuB\":[\"Followers\"],\"nqCai9\":[\"Links\"],\"kBG60d\":[\"Looks like this page is a bit empty.\"],\"5cg++T\":[\"Follow and check back soon for surprise lists!\"],\"BuijXr\":[\"<0>\",[\"0\"],\"</0>\"],\"UrzYP4\":[\"Right now, only invited users can log in\"],\"DGilck\":[\"Already got your invite? Jump in and apply now!\"],\"FATCWC\":[\"New to Relist? Get started!\"],\"TGUAu2\":[\"By continuing, you agree to our <0>Terms of Service</0>, <1>Privacy Policy</1>. You confirm you're 13+.\"],\"v0GDxO\":[\"Enter your text here\"],\"tfDRzk\":[\"Save\"],\"W9FRBT\":[\"Like\"],\"yVFoGP\":[\"Create List\"],\"6eQ0FA\":[\"Image must be at least 150x150px.\"],\"s6QF9r\":[\"Choose your image\"],\"aHlf/t\":[\"500x500px, JPG or PNG, max 4MB\"],\"xHDYLK\":[\"Upload a cover\"],\"VaEPrX\":[\"Join Relist and share your lists.\"],\"Vm9srO\":[\"Currently open only to creator account applications.\"],\"e+RpCP\":[\"Sign up\"],\"/evPCn\":[\"Creator Account Login\"],\"JX4ezt\":[\"Can non-creators use Relist?\"],\"/q7UJE\":[\"Real impact is simple\"],\"JOfh8O\":[\"Turn your picks into a following\"],\"2doGYd\":[\"Inspogirlie\"],\"YzMqiP\":[\"Must-Haves for dorm life\"],\"Jrcjjh\":[\"Cozy Bedding:\"],\"olcKig\":[\"A plush comforter and extra pillows for a comfy night's sleep.\"],\"GPzBEq\":[\"Desk Essentials\"],\"CmxV+G\":[\"Target is probably your bf.\"],\"LKwNpo\":[\"Portable Speaker\"],\"GHHXT0\":[\"For study breaks or casual dorm hangouts\"],\"V/D+9x\":[\"List Lab\"],\"vQ65+r\":[\"Stop-by-Stop Café Journey\"],\"Zzo4uk\":[\"8:00 AM at Sunrise Café\"],\"9DUK98\":[\"Start your day with a creamy latte and a freshly baked croissant\"],\"c+8j4b\":[\"12:30 PM at Artisan Coffee House\"],\"eVVHMA\":[\"Enjoy a flat white with a hearty avocado toast.\"],\"tEpROq\":[\"3:00 PM at Brew & Bloom\"],\"LrHd3m\":[\"Sip on a cold brew while indulging in a lavender scone\"],\"mPs/O3\":[\"Campus People\"],\"x3KB6d\":[\"Notes for Campus Fest\"],\"UQjid/\":[\"Essentials\"],\"nYR44Q\":[\"Sunscreen, water bottle, and comfy shoes.\"],\"LeaWVj\":[\"Extras\"],\"uqajLC\":[\"A light jacket for when it gets chilly.\"],\"F7HoVk\":[\"Pro tip:\"],\"O1b0Cr\":[\"Bring a blanket to claim your spot early.\"],\"4oJHLS\":[\"Travel Lover\"],\"OWjR7r\":[\"Travel budgeting checklist\"],\"Ir1CN/\":[\"Set your total budget\"],\"c2eiO8\":[\"Break it down into categories: flights, accommodation, food, activities, and extras.\"],\"SZ6k2f\":[\"Hunt for Deals\"],\"TLPUAN\":[\"Use flight comparison tools like Skyscanner or Google Flights\"],\"QdUQ+N\":[\"Transportation Hacks\"],\"B1JN8A\":[\"Book accommodations early through budget-friendly platforms like Hostelworld or Airbnb.\"],\"qEIbYn\":[\"Gossip Pie\"],\"jUlzvp\":[\"Binge based on your mood\"],\"0U3C98\":[\"Relaxing: Moana\"],\"CXBJE9\":[\"Moana / A heartwarming and adventurous animated film with stunning visuals and beautiful music that will soothe your soul.\"],\"c3jXdn\":[\"Tense: Tenet\"],\"5Al7Bs\":[\"Directed by Christopher Nolan, this mind-bending action film with time inversion will keep you on the edge of your seat.\"],\"mRxtxe\":[\"Fun: Dumb and Dumber\"],\"xGoWE/\":[\"A classic comedy about two clueless friends on a road trip full of ridiculous and laugh-out-loud situations.\"],\"SMFKok\":[\"iTech\"],\"PdTFHs\":[\"How I curate my digital space\"],\"624Mtx\":[\"Notion\"],\"ba9xfk\":[\"Manage your thoughts and ideas in one single place.\"],\"w9Xmw/\":[\"Google Drive\"],\"Q2veSe\":[\"The OG and it's so essential.\"],\"yOaaWk\":[\"ChatGPT\"],\"DfSZt9\":[\"Greatly improved my work!\"],\"kPUYC2\":[\"Good Habbits\"],\"gS0YFj\":[\"Atomic Habits - My thoughts\"],\"QBV5Dh\":[\"Start Small to Achieve Big\"],\"x3x8gi\":[\"Pressure is the biggest enemy. Don't let it control your momentum.\"],\"nkJmeJ\":[\"Build Systems, Not Just Goals.\"],\"Pzs/Wu\":[\"This reminds me that building team is more important than the actual product.\"],\"hVijR6\":[\"Embrace Your New Identity\"],\"n+De4p\":[\"I need to work on this 100%\"],\"iEfYGf\":[\"Pica Pica\"],\"sPzKc0\":[\"My best protein shake recipe\"],\"rkUWz9\":[\"Step 1: Gather Your Ingredients\"],\"BLAQSh\":[\"1 scoop of your favorite protein powder (vanilla or chocolate works great!)\"],\"HJwOp1\":[\"Step 2: Blend the Base\"],\"8VH0aE\":[\"Add the almond milk, protein powder, and banana to your blender.\"],\"OOBT5G\":[\"Step 3: Add the Extras\"],\"XyYukS\":[\"Drop in the peanut butter, chia seeds, or flaxseeds for added nutrients.\"],\"G9hWXm\":[\"Jacob PHD\"],\"LGdu/V\":[\"Reasons I swipe left\"],\"/DEtrq\":[\"The Fish Pic\"],\"K5eTbU\":[\"Congrats on catching that trout, but it's not winning you dates.\"],\"IiU328\":[\"No Bio, No Effort\"],\"3KK42j\":[\"If you can't write a sentence, how are you going to hold a conversation?\"],\"jSx6Fj\":[\"Mirror Selfies with a Dirty Room\"],\"LZQmU/\":[\"If you can't clean your room, what's your life looking like?\"],\"LB7eBr\":[\"Tutorials\"],\"/lDBHm\":[\"FAQ\"],\"pGWSZE\":[\"Join the Team\"],\"XyA7q/\":[\"Press kit\"],\"YcfUZ9\":[\"Contact Us\"],\"hKIgWM\":[\"Collection Notice\"],\"4VQIM2\":[\"Login Error - Creator Account Not Found\"],\"ie/LhK\":[\"Only approved and verified “Creator Accounts” can log in and create lists. If you have any questions, please contact us.\"],\"bJ55aX\":[\"Please login again\"],\"sYV3p0\":[\"Please login to do this action\"],\"Pdjj9s\":[\"The app is outdated, please login again\"]}")as Messages;
