+++ src/locales/zh-TW/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"6YtxFj\":[\"姓名\"],\"S5Bu+B\":[\"輸入你的姓名\"],\"7sNhEz\":[\"用戶帳號\"],\"zmL8Lp\":[\"輸入你的用戶帳號\"],\"E+oKCO\":[\"個人自介\"],\"O8cFJC\":[\"輸入你的個人自介\"],\"yjVihA\":[\"個人檔案頭貼\"],\"XpjWSd\":[\"自訂連結\"],\"dn+x7E\":[\"輸入網址\"],\"MPPZ54\":[\"輸入帳號\"],\"vERlcd\":[\"個人檔案\"],\"ZCYobq\":[\"連結\"],\"O2GxA2\":[\"編輯個人檔案與帳號\"],\"FxVG/l\":[\"已複製到剪貼簿\"],\"he3ygx\":[\"複製\"],\"ZopSbj\":[\"最近更新\"],\"4Q5LRX\":[\"創單自\"],\"zZOZj7\":[\"喜歡\"],\"tuhghE\":[\"編輯名單\"],\"fOKqw/\":[\"新增靈感\"],\"VSniCT\":[\"名單封面\"],\"ZXeH4g\":[\"靈感名單\"],\"MCLjtL\":[\"編輯名單封面\"],\"VsDFHC\":[\"新增的靈感會在這裡，丟個靈感進來吧！\"],\"IrqaKp\":[\"點擊靈感可編輯，拖曳左側三條線可排序。\"],\"rhRxkK\":[\"編輯名單\"],\"V22EiC\":[\"儲存新排序\"],\"b3p7OM\":[\"刪掉名單標題後，連所有靈感都會消失！\"],\"KMfg9q\":[\"永久刪除整個名單及所有靈感，這將無法復原！\"],\"THIBos\":[\"刪除名單與所有靈感\"],\"dEgA5A\":[\"取消\"],\"P0rbCt\":[\"封面圖片\"],\"gsjH+v\":[\"欸，標題不能留空喔！\"],\"E93Wuj\":[\"名單都得有個標題，趕快填一下吧！\"],\"dBg1Gg\":[\"名單標題字數太長啦！\"],\"J/tFrO\":[\"標題過長，請將字數縮短至 \",[\"TITLE_MAX_LENGTH\"],\" 個字元以內。\"],\"l6w5x5\":[\"標題描述內容超出限制了！\"],\"F6+KZb\":[\"字數有點爆表，請減到 \",[\"DESC_MAX_LENGTH\"],\" 個字元以內。\"],\"JgzDWT\":[\"在這輸入名單標題\"],\"w/K9b6\":[\"描述標題\"],\"KHDCQm\":[\"連結網頁\"],\"nllOOr\":[\"編輯名單分類\"],\"gbxG1z\":[\"名單主題\"],\"FVbPOI\":[\"選擇一項接近此份名單的分類\"],\"hXzOVo\":[\"下一步\"],\"edl1Bx\":[\"取消後，資料就飛走囉！\"],\"5ttevj\":[\"如果取消，所填的內容都會消失\"],\"PsVVX9\":[\"取消編輯\"],\"hoaavJ\":[\"繼續編輯\"],\"AOvWO2\":[\"建立靈感名單\"],\"/IX/7x\":[\"其他\"],\"rBfbli\":[\"生活風格\"],\"lWxAUo\":[\"美食\"],\"wFuUuH\":[\"文化\"],\"lj2WeU\":[\"旅遊\"],\"x7ZLZ3\":[\"娛樂\"],\"kAgdQQ\":[\"數位科技\"],\"FgFxl9\":[\"個人成長\"],\"bLbRrJ\":[\"健康與健身\"],\"uyJsf6\":[\"關於我們\"],\"4Y5H+g\":[\"使用者條款\"],\"rjGI/Q\":[\"隱私權保護政策\"],\"ab3OdJ\":[\"Relist 在手機畫面運作更順暢，使用行動裝置以取得更好的使用體驗。\"],\"/fOrgx\":[\"<0>你的生活精選，比誰都有料</0><1><2>下個內容浪潮</2><3/></1>\"],\"sfb9Ng\":[\"<0>靈感都在我的名單裡</0><1>歡迎受邀朋友申請帳號！</1>\"],\"cpE88+\":[\"申請帳號\"],\"5lWFkC\":[\"登入\"],\"NuKR0h\":[\"其他\"],\"jAN+Se\":[\"這 vibe 我可以\"],\"8YtR7h\":[\"好想吃吃喝喝\"],\"DLW/m0\":[\"有點溫度的東西\"],\"o2i9EJ\":[\"出去玩耍的私筆記\"],\"MlvbpW\":[\"耳機一戴，世界走開\"],\"2GbPea\":[\"正在生成的浪潮\"],\"3GYIUI\":[\"慢慢懂的一些事\"],\"YeTwPt\":[\"練出好狀態\"],\"IUd4br\":[\"喜歡的都在這\"],\"EiP0f+\":[\"<0>最近更新</0>\"],\"V1fa9u\":[\"展開更多\"],\"Zs3bDD\":[\"這些靈感名單超有感\"],\"ZJ0qYg\":[\"每週精選更新\"],\"rP2JNC\":[\"哎呀！網頁又在耍脾氣啦！\"],\"hYyEgm\":[\"<0>Relist 在手機畫面運作更順暢</0><1>使用行動裝置以取得更好的使用體驗</1>\"],\"tWA9nu\":[\"靈感\"],\"AJoiN9\":[\"確定要刪除此靈感嗎？\"],\"KCgg1o\":[\"刪除靈感後將無法復原！\"],\"N6GBcC\":[\"確定刪除\"],\"3YNSm9\":[\"靈感都得有個標題，趕快填一下吧！\"],\"ClX0MT\":[\"靈感標題字數太長啦！\"],\"Z/FWct\":[\"在這輸入靈感標題\"],\"G4krEw\":[\"描述靈感\"],\"gMn81o\":[\"名單已經存到你的個人專頁了，但剛寫的靈感內容會消失，確定要取消嗎？\"],\"jf150A\":[\"編輯靈感\"],\"D9WD3n\":[\"新靈感\"],\"DPfwMq\":[\"完成\"],\"0ESV18\":[\"生活風格\"],\"h16FyT\":[\"名單\"],\"VQmsFR\":[\"選擇您的語言\"],\"rtZ04n\":[\"選擇您喜愛的語言，讓使用體驗更順暢。\"],\"if5Kxg\":[\"台灣\"],\"Mqy/Zy\":[\"美國\"],\"E6OqPk\":[\"選擇您的所在地\"],\"FsmHKW\":[\"選擇您的所在地，讓使用體驗更個人化。\"],\"iZJxvU\":[\"偏好設定\"],\"AXTJAW\":[\"選擇您的語言\"],\"FFy8Oz\":[\"選擇您的所在地\"],\"DNeL+Q\":[\"關於 Relist\"],\"8jGB5r\":[\"三分鐘上手教學\"],\"RJ5xTx\":[\"Relist 官網\"],\"3fMxeH\":[\"追蹤 Relist Threads\"],\"YbVeAM\":[\"加入 Discord 與我們一起優化 Relist\"],\"bucbEL\":[\"給予意見或內容檢舉\"],\"2OUiOQ\":[\"關於隱私政策與使用條款\"],\"Bhgd0l\":[\"聯繫我們\"],\"n1ekoW\":[\"登入\"],\"vzX5FB\":[\"刪除帳號\"],\"bHYIks\":[\"登出\"],\"VdL8Lz\":[\"歡迎來到 Relist！\"],\"zaNICH\":[\"這裡是你分享想法、建立口袋名單的地方。歡迎與我們分享你的使用經驗，讓我們一起把 Relist 做得更棒！\"],\"OEtKMx\":[\"分享你的意見\"],\"EjIvar\":[\"設定中心\"],\"k63iCK\":[\"登入已過期，請重新登入\"],\"y6sq5j\":[\"正追蹤\"],\"MKEPCY\":[\"追蹤\"],\"x5LEuB\":[\"粉絲\"],\"nqCai9\":[\"個連結\"],\"kBG60d\":[\"看起來這個頁面有點空。\"],\"5cg++T\":[\"先追蹤起來，隨時回來看看吧！\"],\"BuijXr\":[\"<0>\",[\"0\"],\"</0>\"],\"FATCWC\":[\"第一次登入 RELIST ？立即申請\"],\"TGUAu2\":[\"登入或註冊即表示您已閱讀並同意我們的<0>《使用者條款》</0>與<1>《隱私權保護政策》</1>，並確認您已年滿 13 歲。\"],\"v0GDxO\":[\"請輸入內容\"],\"tfDRzk\":[\"儲存\"],\"W9FRBT\":[\"喜歡\"],\"yVFoGP\":[\"建立名單\"],\"6eQ0FA\":[\"圖片必須至少 150x150 像素\"],\"s6QF9r\":[\"選擇圖片\"],\"aHlf/t\":[\"500x500像素，限 JPG 或 PNG，最大4MB\"],\"xHDYLK\":[\"封面圖片\"],\"UrzYP4\":[\"現在僅開放受邀請的用戶登入體驗\"],\"DGilck\":[\"收到邀請了嗎？現在就去申請加入！\"],\"VaEPrX\":[\"加入 Relist，分享靈感！\"],\"Vm9srO\":[\"目前僅開放創作者申請帳號\"],\"e+RpCP\":[\"快速申請\"],\"/evPCn\":[\"創作者帳號登入\"],\"JX4ezt\":[\"非創作者也能加入嗎？\"],\"/q7UJE\":[\"簡單打造影響力\"],\"JOfh8O\":[\"你的精選就能吸引懂你的人\"],\"2doGYd\":[\"小資妹\"],\"YzMqiP\":[\"大學宿舍日常必備清單\"],\"Jrcjjh\":[\"舒服的海綿床\"],\"olcKig\":[\"我發現好在床上的時間比想像的多，有一個舒服的床真的舒適許多。\"],\"GPzBEq\":[\"桌上收納盒\"],\"CmxV+G\":[\"桌上的收納盒超重要，讓桌上乾淨又一目暸然。\"],\"LKwNpo\":[\"無線隨身音響\"],\"GHHXT0\":[\"耳機是必備，但無線音響讓聽音樂更自由舒適，超愛的！\"],\"V/D+9x\":[\"List Lab\"],\"vQ65+r\":[\"我的週末咖啡日記\"],\"Zzo4uk\":[\"早上 8:00 小香咖啡\"],\"9DUK98\":[\"我家附近開最早的咖啡！\"],\"c+8j4b\":[\"中午 12:30 爵士咖啡館\"],\"eVVHMA\":[\"最近週末迷上爵士音樂，這裡真的超放鬆的！\"],\"tEpROq\":[\"下午 3:00 Brew & Bloom\"],\"LrHd3m\":[\"完全是衝著他們的點心而來！\"],\"mPs/O3\":[\"大學生秘技\"],\"x3KB6d\":[\"筆記：去音樂祭前一定要看\"],\"UQjid/\":[\"超基本配備\"],\"nYR44Q\":[\"水壺與輕便的鞋子非常重要！防曬乳千萬不要忘了！\"],\"LeaWVj\":[\"多帶一件\"],\"uqajLC\":[\"Tshirt 最好多帶一件，薄外套絕對不能少！\"],\"F7HoVk\":[\"超重要秘技\"],\"O1b0Cr\":[\"一定要帶地墊與手提袋，佔地的好武器！\"],\"4oJHLS\":[\"無聊就輸了\"],\"OWjR7r\":[\"小資旅遊的好方法\"],\"Ir1CN/\":[\"決定好可花的總金額\"],\"c2eiO8\":[\"車票、機票與飯店幾乎是固定的費用，抓好總金額就對了！\"],\"SZ6k2f\":[\"只搜特惠\"],\"TLPUAN\":[\"使用訂票時，打開篩選特惠，先專注特惠價格。\"],\"QdUQ+N\":[\"訂票快狠準\"],\"B1JN8A\":[\"不是越早訂越便宜，有時反而是越晚訂越優惠！\"],\"qEIbYn\":[\"八卦生活\"],\"jUlzvp\":[\"按心情的追劇法\"],\"0U3C98\":[\"放鬆: 海洋奇緣\"],\"CXBJE9\":[\"一部充滿溫暖與冒險的動畫電影，畫面美麗，音樂動人，帶來滿滿的療癒感。\"],\"c3jXdn\":[\"緊張: 天能 \"],\"5Al7Bs\":[\"克里斯多福諾蘭執導的燒腦動作片，充滿時間逆轉的緊湊劇情，讓你全程屏息。\"],\"mRxtxe\":[\"有趣: 阿呆與阿瓜\"],\"xGoWE/\":[\"兩位完全笨到極致的好朋友，展開一場荒唐的公路旅行，充滿令人哭笑不得的鬧劇！這部經典絕對讓你笑到停不下來。\"],\"SMFKok\":[\"讀讀科技\"],\"PdTFHs\":[\"懶人提升效率的APP\"],\"624Mtx\":[\"Notion\"],\"ba9xfk\":[\"整合筆記與項目管理，效率爆表。一定要去下載!\"],\"w9Xmw/\":[\"Google Drive\"],\"Q2veSe\":[\"最近我連健身運動都用GD! 有興趣的可以跟我說\"],\"yOaaWk\":[\"ChatGPT\"],\"DfSZt9\":[\"AI真的省去我50%的工作時間，而且效率還超越\"],\"kPUYC2\":[\"上班中的妞\"],\"gS0YFj\":[\"《原子習慣》讀到的智慧…\"],\"QBV5Dh\":[\"從小開始\"],\"x3x8gi\":[\"一口氣想做大事，往往最後一事無成啊！\"],\"nkJmeJ\":[\"打造一個系統，而非結果！\"],\"Pzs/Wu\":[\"這個超有感的！過去我總是想要有一個厲害的結果。\"],\"hVijR6\":[\"接受新的身份\"],\"n+De4p\":[\"這思考非常特別，有時候自己對自己的價值任何能夠改變事情的開始與結果。\"],\"iEfYGf\":[\"Pica Pica\"],\"sPzKc0\":[\"高蛋白飲料~原來超簡單做的\"],\"rkUWz9\":[\"步驟一 ：把所有材料準備好\"],\"BLAQSh\":[\"先把東西都放在一起準備好，因為超簡單的！\"],\"HJwOp1\":[\"步驟二: 先把基礎的倒進來\"],\"8VH0aE\":[\"把牛奶與高蛋白先倒在一起，如此就已經做對一半！\"],\"OOBT5G\":[\"步驟三：加入增肌配方\"],\"XyYukS\":[\"堅果是超棒的配方，馬上讓你有飽足感又很可口好喝！\"],\"G9hWXm\":[\"末日研究生\"],\"LGdu/V\":[\"絕對左甩的交友檔案\"],\"/DEtrq\":[\"太詐騙的照片\"],\"K5eTbU\":[\"老實說...我每次都覺得這麼假的照片也敢放上來，當作大家是瞎子嗎？\"],\"IiU328\":[\"沒有編輯的自介文字\"],\"3KK42j\":[\"有些人的自介人格分裂，然後又錯字連篇，我真心無法違背自己良心右甩。\"],\"jSx6Fj\":[\"自拍照裡的房間亂到像核爆\"],\"LZQmU/\":[\"這是我的問題，但我眼睛真的很利！哈！\"],\"LB7eBr\":[\"快速教學\"],\"/lDBHm\":[\"問與答\"],\"pGWSZE\":[\"加入團隊\"],\"XyA7q/\":[\"媒體報導\"],\"YcfUZ9\":[\"聯繫我們\"],\"hKIgWM\":[\"額外公告\"],\"4VQIM2\":[\"登入錯誤 - 查無此創作者\"],\"ie/LhK\":[\"目前僅限已申請並被認證的「創作者帳號」登入並創作名單。如有疑問，請聯繫我們。\"],\"bJ55aX\":[\"請重新登入\"],\"sYV3p0\":[\"請先登入以進行此操作\"],\"Pdjj9s\":[\"網頁更新，請重新登入\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"6YtxFj\":[\"姓名\"],\"S5Bu+B\":[\"輸入你的姓名\"],\"7sNhEz\":[\"用戶帳號\"],\"zmL8Lp\":[\"輸入你的用戶帳號\"],\"E+oKCO\":[\"個人自介\"],\"O8cFJC\":[\"輸入你的個人自介\"],\"yjVihA\":[\"個人檔案頭貼\"],\"XpjWSd\":[\"自訂連結\"],\"dn+x7E\":[\"輸入網址\"],\"MPPZ54\":[\"輸入帳號\"],\"vERlcd\":[\"個人檔案\"],\"ZCYobq\":[\"連結\"],\"O2GxA2\":[\"編輯個人檔案與帳號\"],\"FxVG/l\":[\"已複製到剪貼簿\"],\"he3ygx\":[\"複製\"],\"ZopSbj\":[\"最近更新\"],\"4Q5LRX\":[\"創單自\"],\"zZOZj7\":[\"喜歡\"],\"tuhghE\":[\"編輯名單\"],\"fOKqw/\":[\"新增靈感\"],\"VSniCT\":[\"名單封面\"],\"ZXeH4g\":[\"靈感名單\"],\"MCLjtL\":[\"編輯名單封面\"],\"VsDFHC\":[\"新增的靈感會在這裡，丟個靈感進來吧！\"],\"IrqaKp\":[\"點擊靈感可編輯，拖曳左側三條線可排序。\"],\"rhRxkK\":[\"編輯名單\"],\"V22EiC\":[\"儲存新排序\"],\"b3p7OM\":[\"刪掉名單標題後，連所有靈感都會消失！\"],\"KMfg9q\":[\"永久刪除整個名單及所有靈感，這將無法復原！\"],\"THIBos\":[\"刪除名單與所有靈感\"],\"dEgA5A\":[\"取消\"],\"P0rbCt\":[\"封面圖片\"],\"gsjH+v\":[\"欸，標題不能留空喔！\"],\"E93Wuj\":[\"名單都得有個標題，趕快填一下吧！\"],\"dBg1Gg\":[\"名單標題字數太長啦！\"],\"J/tFrO\":[\"標題過長，請將字數縮短至 \",[\"TITLE_MAX_LENGTH\"],\" 個字元以內。\"],\"l6w5x5\":[\"標題描述內容超出限制了！\"],\"F6+KZb\":[\"字數有點爆表，請減到 \",[\"DESC_MAX_LENGTH\"],\" 個字元以內。\"],\"JgzDWT\":[\"在這輸入名單標題\"],\"w/K9b6\":[\"描述標題\"],\"KHDCQm\":[\"連結網頁\"],\"nllOOr\":[\"編輯名單分類\"],\"gbxG1z\":[\"名單主題\"],\"FVbPOI\":[\"選擇一項接近此份名單的分類\"],\"hXzOVo\":[\"下一步\"],\"edl1Bx\":[\"取消後，資料就飛走囉！\"],\"5ttevj\":[\"如果取消，所填的內容都會消失\"],\"PsVVX9\":[\"取消編輯\"],\"hoaavJ\":[\"繼續編輯\"],\"AOvWO2\":[\"建立靈感名單\"],\"/IX/7x\":[\"其他\"],\"rBfbli\":[\"生活風格\"],\"lWxAUo\":[\"美食\"],\"wFuUuH\":[\"文化\"],\"lj2WeU\":[\"旅遊\"],\"x7ZLZ3\":[\"娛樂\"],\"kAgdQQ\":[\"數位科技\"],\"FgFxl9\":[\"個人成長\"],\"bLbRrJ\":[\"健康與健身\"],\"uyJsf6\":[\"關於我們\"],\"4Y5H+g\":[\"使用者條款\"],\"rjGI/Q\":[\"隱私權保護政策\"],\"ab3OdJ\":[\"Relist 在手機畫面運作更順暢，使用行動裝置以取得更好的使用體驗。\"],\"/fOrgx\":[\"<0>你的生活精選，比誰都有料</0><1><2>下個內容浪潮</2><3/></1>\"],\"sfb9Ng\":[\"<0>靈感都在我的名單裡</0><1>歡迎受邀朋友申請帳號！</1>\"],\"cpE88+\":[\"申請帳號\"],\"5lWFkC\":[\"登入\"],\"NuKR0h\":[\"其他\"],\"jAN+Se\":[\"這 vibe 我可以\"],\"8YtR7h\":[\"好想吃吃喝喝\"],\"DLW/m0\":[\"有點溫度的東西\"],\"o2i9EJ\":[\"出去玩耍的私筆記\"],\"MlvbpW\":[\"耳機一戴，世界走開\"],\"2GbPea\":[\"正在生成的浪潮\"],\"3GYIUI\":[\"慢慢懂的一些事\"],\"YeTwPt\":[\"練出好狀態\"],\"IUd4br\":[\"喜歡的都在這\"],\"EiP0f+\":[\"<0>最近更新</0>\"],\"V1fa9u\":[\"展開更多\"],\"Zs3bDD\":[\"這些靈感名單超有感\"],\"ZJ0qYg\":[\"每週精選更新\"],\"rP2JNC\":[\"哎呀！網頁又在耍脾氣啦！\"],\"hYyEgm\":[\"<0>Relist 在手機畫面運作更順暢</0><1>使用行動裝置以取得更好的使用體驗</1>\"],\"tWA9nu\":[\"靈感\"],\"AJoiN9\":[\"確定要刪除此靈感嗎？\"],\"KCgg1o\":[\"刪除靈感後將無法復原！\"],\"N6GBcC\":[\"確定刪除\"],\"3YNSm9\":[\"靈感都得有個標題，趕快填一下吧！\"],\"ClX0MT\":[\"靈感標題字數太長啦！\"],\"Z/FWct\":[\"在這輸入靈感標題\"],\"G4krEw\":[\"描述靈感\"],\"gMn81o\":[\"名單已經存到你的個人專頁了，但剛寫的靈感內容會消失，確定要取消嗎？\"],\"jf150A\":[\"編輯靈感\"],\"D9WD3n\":[\"新靈感\"],\"DPfwMq\":[\"完成\"],\"0ESV18\":[\"生活風格\"],\"h16FyT\":[\"名單\"],\"VQmsFR\":[\"選擇您的語言\"],\"rtZ04n\":[\"選擇您喜愛的語言，讓使用體驗更順暢。\"],\"if5Kxg\":[\"台灣\"],\"Mqy/Zy\":[\"美國\"],\"E6OqPk\":[\"選擇您的所在地\"],\"FsmHKW\":[\"選擇您的所在地，讓使用體驗更個人化。\"],\"iZJxvU\":[\"偏好設定\"],\"AXTJAW\":[\"選擇您的語言\"],\"FFy8Oz\":[\"選擇您的所在地\"],\"DNeL+Q\":[\"關於 Relist\"],\"8jGB5r\":[\"三分鐘上手教學\"],\"RJ5xTx\":[\"Relist 官網\"],\"3fMxeH\":[\"追蹤 Relist Threads\"],\"YbVeAM\":[\"加入 Discord 與我們一起優化 Relist\"],\"bucbEL\":[\"給予意見或內容檢舉\"],\"2OUiOQ\":[\"關於隱私政策與使用條款\"],\"Bhgd0l\":[\"聯繫我們\"],\"n1ekoW\":[\"登入\"],\"vzX5FB\":[\"刪除帳號\"],\"bHYIks\":[\"登出\"],\"VdL8Lz\":[\"歡迎來到 Relist！\"],\"zaNICH\":[\"這裡是你分享想法、建立口袋名單的地方。歡迎與我們分享你的使用經驗，讓我們一起把 Relist 做得更棒！\"],\"OEtKMx\":[\"分享你的意見\"],\"EjIvar\":[\"設定中心\"],\"k63iCK\":[\"登入已過期，請重新登入\"],\"y6sq5j\":[\"正追蹤\"],\"MKEPCY\":[\"追蹤\"],\"x5LEuB\":[\"粉絲\"],\"nqCai9\":[\"個連結\"],\"kBG60d\":[\"看起來這個頁面有點空。\"],\"5cg++T\":[\"先追蹤起來，隨時回來看看吧！\"],\"BuijXr\":[\"<0>\",[\"0\"],\"</0>\"],\"UrzYP4\":[\"現在僅開放受邀請的用戶登入體驗\"],\"DGilck\":[\"收到邀請了嗎？現在就去申請加入！\"],\"FATCWC\":[\"第一次登入 RELIST ？立即申請\"],\"TGUAu2\":[\"登入或註冊即表示您已閱讀並同意我們的<0>《使用者條款》</0>與<1>《隱私權保護政策》</1>，並確認您已年滿 13 歲。\"],\"v0GDxO\":[\"請輸入內容\"],\"tfDRzk\":[\"儲存\"],\"W9FRBT\":[\"喜歡\"],\"yVFoGP\":[\"建立名單\"],\"6eQ0FA\":[\"圖片必須至少 150x150 像素\"],\"s6QF9r\":[\"選擇圖片\"],\"aHlf/t\":[\"500x500像素，限 JPG 或 PNG，最大4MB\"],\"xHDYLK\":[\"封面圖片\"],\"VaEPrX\":[\"加入 Relist，分享靈感！\"],\"Vm9srO\":[\"目前僅開放創作者申請帳號\"],\"e+RpCP\":[\"快速申請\"],\"/evPCn\":[\"創作者帳號登入\"],\"JX4ezt\":[\"非創作者也能加入嗎？\"],\"/q7UJE\":[\"簡單打造影響力\"],\"JOfh8O\":[\"你的精選就能吸引懂你的人\"],\"2doGYd\":[\"小資妹\"],\"YzMqiP\":[\"大學宿舍日常必備清單\"],\"Jrcjjh\":[\"舒服的海綿床\"],\"olcKig\":[\"我發現好在床上的時間比想像的多，有一個舒服的床真的舒適許多。\"],\"GPzBEq\":[\"桌上收納盒\"],\"CmxV+G\":[\"桌上的收納盒超重要，讓桌上乾淨又一目暸然。\"],\"LKwNpo\":[\"無線隨身音響\"],\"GHHXT0\":[\"耳機是必備，但無線音響讓聽音樂更自由舒適，超愛的！\"],\"V/D+9x\":[\"List Lab\"],\"vQ65+r\":[\"我的週末咖啡日記\"],\"Zzo4uk\":[\"早上 8:00 小香咖啡\"],\"9DUK98\":[\"我家附近開最早的咖啡！\"],\"c+8j4b\":[\"中午 12:30 爵士咖啡館\"],\"eVVHMA\":[\"最近週末迷上爵士音樂，這裡真的超放鬆的！\"],\"tEpROq\":[\"下午 3:00 Brew & Bloom\"],\"LrHd3m\":[\"完全是衝著他們的點心而來！\"],\"mPs/O3\":[\"大學生秘技\"],\"x3KB6d\":[\"筆記：去音樂祭前一定要看\"],\"UQjid/\":[\"超基本配備\"],\"nYR44Q\":[\"水壺與輕便的鞋子非常重要！防曬乳千萬不要忘了！\"],\"LeaWVj\":[\"多帶一件\"],\"uqajLC\":[\"Tshirt 最好多帶一件，薄外套絕對不能少！\"],\"F7HoVk\":[\"超重要秘技\"],\"O1b0Cr\":[\"一定要帶地墊與手提袋，佔地的好武器！\"],\"4oJHLS\":[\"無聊就輸了\"],\"OWjR7r\":[\"小資旅遊的好方法\"],\"Ir1CN/\":[\"決定好可花的總金額\"],\"c2eiO8\":[\"車票、機票與飯店幾乎是固定的費用，抓好總金額就對了！\"],\"SZ6k2f\":[\"只搜特惠\"],\"TLPUAN\":[\"使用訂票時，打開篩選特惠，先專注特惠價格。\"],\"QdUQ+N\":[\"訂票快狠準\"],\"B1JN8A\":[\"不是越早訂越便宜，有時反而是越晚訂越優惠！\"],\"qEIbYn\":[\"八卦生活\"],\"jUlzvp\":[\"按心情的追劇法\"],\"0U3C98\":[\"放鬆: 海洋奇緣\"],\"CXBJE9\":[\"一部充滿溫暖與冒險的動畫電影，畫面美麗，音樂動人，帶來滿滿的療癒感。\"],\"c3jXdn\":[\"緊張: 天能 \"],\"5Al7Bs\":[\"克里斯多福諾蘭執導的燒腦動作片，充滿時間逆轉的緊湊劇情，讓你全程屏息。\"],\"mRxtxe\":[\"有趣: 阿呆與阿瓜\"],\"xGoWE/\":[\"兩位完全笨到極致的好朋友，展開一場荒唐的公路旅行，充滿令人哭笑不得的鬧劇！這部經典絕對讓你笑到停不下來。\"],\"SMFKok\":[\"讀讀科技\"],\"PdTFHs\":[\"懶人提升效率的APP\"],\"624Mtx\":[\"Notion\"],\"ba9xfk\":[\"整合筆記與項目管理，效率爆表。一定要去下載!\"],\"w9Xmw/\":[\"Google Drive\"],\"Q2veSe\":[\"最近我連健身運動都用GD! 有興趣的可以跟我說\"],\"yOaaWk\":[\"ChatGPT\"],\"DfSZt9\":[\"AI真的省去我50%的工作時間，而且效率還超越\"],\"kPUYC2\":[\"上班中的妞\"],\"gS0YFj\":[\"《原子習慣》讀到的智慧…\"],\"QBV5Dh\":[\"從小開始\"],\"x3x8gi\":[\"一口氣想做大事，往往最後一事無成啊！\"],\"nkJmeJ\":[\"打造一個系統，而非結果！\"],\"Pzs/Wu\":[\"這個超有感的！過去我總是想要有一個厲害的結果。\"],\"hVijR6\":[\"接受新的身份\"],\"n+De4p\":[\"這思考非常特別，有時候自己對自己的價值任何能夠改變事情的開始與結果。\"],\"iEfYGf\":[\"Pica Pica\"],\"sPzKc0\":[\"高蛋白飲料~原來超簡單做的\"],\"rkUWz9\":[\"步驟一 ：把所有材料準備好\"],\"BLAQSh\":[\"先把東西都放在一起準備好，因為超簡單的！\"],\"HJwOp1\":[\"步驟二: 先把基礎的倒進來\"],\"8VH0aE\":[\"把牛奶與高蛋白先倒在一起，如此就已經做對一半！\"],\"OOBT5G\":[\"步驟三：加入增肌配方\"],\"XyYukS\":[\"堅果是超棒的配方，馬上讓你有飽足感又很可口好喝！\"],\"G9hWXm\":[\"末日研究生\"],\"LGdu/V\":[\"絕對左甩的交友檔案\"],\"/DEtrq\":[\"太詐騙的照片\"],\"K5eTbU\":[\"老實說...我每次都覺得這麼假的照片也敢放上來，當作大家是瞎子嗎？\"],\"IiU328\":[\"沒有編輯的自介文字\"],\"3KK42j\":[\"有些人的自介人格分裂，然後又錯字連篇，我真心無法違背自己良心右甩。\"],\"jSx6Fj\":[\"自拍照裡的房間亂到像核爆\"],\"LZQmU/\":[\"這是我的問題，但我眼睛真的很利！哈！\"],\"LB7eBr\":[\"快速教學\"],\"/lDBHm\":[\"問與答\"],\"pGWSZE\":[\"加入團隊\"],\"XyA7q/\":[\"媒體報導\"],\"YcfUZ9\":[\"聯繫我們\"],\"hKIgWM\":[\"額外公告\"],\"4VQIM2\":[\"登入錯誤 - 查無此創作者\"],\"ie/LhK\":[\"目前僅限已申請並被認證的「創作者帳號」登入並創作名單。如有疑問，請聯繫我們。\"],\"bJ55aX\":[\"請重新登入\"],\"sYV3p0\":[\"請先登入以進行此操作\"],\"Pdjj9s\":[\"網頁更新，請重新登入\"]}")as Messages;
