+++ src/types/Discovery/index.ts
@@ -24,14 +24,3 @@ export interface LatestList {
 export interface LatestListGroup {
   [key: string]: LatestList[];
 }
-
-export interface LoginDrawerProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onLogin: (success: boolean) => void;
-  onError: () => void;
-}
-export interface LoginInfo {
-  accessToken: string;
-  user: User;
-}
